+++ project/app/src/androidTestGms/java/org/owntracks/android/ui/map/GMSMapActivityTests.kt
@@ -1,8 +1,8 @@
 package org.owntracks.android.ui.map
 
 import android.Manifest
+import android.content.Context
 import androidx.preference.PreferenceManager
-import androidx.test.espresso.Espresso
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
 import androidx.test.platform.app.InstrumentationRegistry
@@ -10,12 +10,9 @@ import com.adevinta.android.barista.assertion.BaristaDrawerAssertions.assertDraw
 import com.adevinta.android.barista.assertion.BaristaEnabledAssertions.assertEnabled
 import com.adevinta.android.barista.assertion.BaristaVisibilityAssertions.assertDisplayed
 import com.adevinta.android.barista.assertion.BaristaVisibilityAssertions.assertNotExist
-import com.adevinta.android.barista.interaction.BaristaClickInteractions.clickBack
 import com.adevinta.android.barista.interaction.BaristaClickInteractions.clickOn
 import com.adevinta.android.barista.interaction.BaristaDialogInteractions.clickDialogNegativeButton
-import com.adevinta.android.barista.interaction.BaristaDialogInteractions.clickDialogPositiveButton
 import com.adevinta.android.barista.interaction.BaristaDrawerInteractions.openDrawer
-import com.adevinta.android.barista.interaction.BaristaEditTextInteractions.writeTo
 import com.adevinta.android.barista.interaction.PermissionGranter
 import org.junit.FixMethodOrder
 import org.junit.Test
@@ -23,9 +20,9 @@ import org.junit.runner.RunWith
 import org.junit.runners.MethodSorters
 import org.owntracks.android.R
 import org.owntracks.android.support.Preferences
+import org.owntracks.android.support.preferences.SharedPreferencesStore
 import org.owntracks.android.testutils.TestWithAnActivity
 import org.owntracks.android.testutils.setNotFirstStartPreferences
-import org.owntracks.android.ui.clickOnAndWait
 
 
 @LargeTest
@@ -87,29 +84,6 @@ class GMSMapActivityTests : TestWithAnActivity<MapActivity>(MapActivity::class.j
     }
 
     @Test
-    fun enablingOSMMapSwitchesFromGMSMapToOSMMap() {
-        setNotFirstStartPreferences()
-        launchActivity()
-        PermissionGranter.allowPermissionsIfNeeded(Manifest.permission.ACCESS_FINE_LOCATION)
-        assertDisplayed(R.id.google_map_view)
-
-        openDrawer()
-        clickOnAndWait(R.string.title_activity_preferences)
-        clickOnAndWait(R.string.configurationManagement)
-        Espresso.openActionBarOverflowOrOptionsMenu(baristaRule.activityTestRule.activity)
-        clickOn(R.string.preferencesEditor)
-        writeTo(
-            R.id.inputKey,
-            baristaRule.activityTestRule.activity.getString(R.string.preferenceKeyExperimentalFeatures)
-        )
-        writeTo(R.id.inputValue, Preferences.EXPERIMENTAL_FEATURE_USE_OSM_MAP)
-        clickDialogPositiveButton()
-        clickBack()
-        clickBack()
-        assertDisplayed(R.id.osm_map_view)
-    }
-
-    @Test
     fun modeButtonOnMapActivityCyclesThroughModes() {
         setNotFirstStartPreferences()
         launchActivity()
@@ -156,5 +130,35 @@ class GMSMapActivityTests : TestWithAnActivity<MapActivity>(MapActivity::class.j
                 .executeShellCommand("settings put secure location_mode 3")
                 .close()
         }
+    }
+
+    @Test
+    fun mapCanSwitchLayerStyleToOsmAndBackAgain() {
+        setNotFirstStartPreferences()
+        launchActivity()
+        PermissionGranter.allowPermissionsIfNeeded(Manifest.permission.ACCESS_FINE_LOCATION)
+        assertDisplayed(R.id.google_map_view)
+        clickOn(R.id.menuMapLayer)
+        clickOn(R.string.menuMapLayerOpenStreetMap)
+        assertDisplayed(R.id.osm_map_view)
+        clickOn(R.id.menuMapLayer)
+        clickOn(R.string.menuMapLayerGoogleMapSatellite)
+        assertDisplayed(R.id.google_map_view)
+    }
+
+    @Test
+    fun mapStartsOnOSMMapIfPreferenceIsSelected() {
+        setNotFirstStartPreferences()
+        val context = InstrumentationRegistry.getInstrumentation().targetContext
+        context.getSharedPreferences(SharedPreferencesStore.FILENAME_PRIVATE, Context.MODE_PRIVATE)
+            .edit()
+            .putString(
+                context.getString(R.string.preferenceKeyMapLayerStyle),
+                MapLayerStyle.OpenStreetMapNormal.name
+            )
+            .apply()
+        launchActivity()
+        PermissionGranter.allowPermissionsIfNeeded(Manifest.permission.ACCESS_FINE_LOCATION)
+        assertDisplayed(R.id.osm_map_view)
     }
 }