+++ project/app/src/gms/java/org/owntracks/android/ui/map/MapFragmentFactory.kt
@@ -4,7 +4,6 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentFactory
 import org.owntracks.android.support.ContactImageBindingAdapter
 import org.owntracks.android.support.Preferences
-import org.owntracks.android.ui.map.osm.OSMMapFragment
 import timber.log.Timber
 import javax.inject.Inject
 
@@ -22,13 +21,10 @@ class MapFragmentFactory @Inject constructor(
 ) : FragmentFactory() {
     override fun instantiate(classLoader: ClassLoader, className: String): Fragment {
         Timber.d("Instantiating Fragment for $className")
-
         return if (MapFragment::class.java.isAssignableFrom(classLoader.loadClass(className))) {
-            if (preferences.experimentalFeatures.contains(Preferences.EXPERIMENTAL_FEATURE_USE_OSM_MAP)) {
-                OSMMapFragment(preferences, contactImageBindingAdapter)
-            } else {
-                GoogleMapFragment(preferences, contactImageBindingAdapter)
-            }
+            preferences.mapLayerStyle.getFragmentClass()
+                .getConstructor(Preferences::class.java, ContactImageBindingAdapter::class.java)
+                .newInstance(preferences, contactImageBindingAdapter)
         } else {
             super.instantiate(classLoader, className)
         }