+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New features
 
+* OpenStreetMap now available as a map layer (!). Layer style toggle switch between the different Google Maps layer styles (Default, Hybrid, Satellite, Terrain), OpenStreetMap and Wikimedia. (#1181)
 * Use new Google Maps renderer on gms flavour: https://developers.google.com/maps/documentation/android-sdk/renderer
 * OSM map now uses a blue dot and arrow for current location (#1078)
 * Add preference to draw regions on the map (#1068)